requirements.txt
backports.tarfile==1.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
idna==3.7
importlib_metadata==8.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
keyring==25.2.1
keyrings.alt==5.0.1
more-itertools==10.3.0
pyicloud==1.0.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.32.3
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
zipp==3.19.2
